refactor: isolated component
