+++ packages/shared/src/components/tooltips/notifications/index.ts
@@ -1,0 +1,1 @@
+export * from './NotificationPreferenceMenu';
